Fish Food / Game Over.
Script added called Food Boundaries to spawn food in random positions and previous updates were made to the Food Controller script. This update takes the function from Food Boundaries to spawn another food pellet around the tank if the player touches/eats it. RigidBody script updated so that the player will disappear if it makes contact with the enemy fish (real fish) and will display "Game Over!".  The Score Manager script was updated to add this function called game over which is then used in RigidBody.